Create partition cleaning utility script
Issues:
Fixes #464

Problem:
In the event the agent to get out of sync with the neutron database,
it is helpful to clear all objects from the tenant partition without
having to navigate to the BIG-IP and delete manually or having to
revert to a saved config. This script will take a folder and BIG-IP
credentials and purge a folder of all its contents.

Analysis:
Adding a script that will parse the configuration file for credentials
and use those to connect to the BIG-IP and remove all objects from
the tenant partition.

Tests:
Populate the tenant folder by creating an LTM service.  Run
the script to remove all objects.